fixing bugs with tests
